+++ lib/sycamore/sycamore/transforms/summarize_images.py
@@ -3,7 +3,7 @@ from typing import Optional
 from PIL import Image
 
 from sycamore.data import Document, Element
-from sycamore.llms import LLM
+from sycamore.llms.llms import LLM
 from sycamore.llms.openai import OpenAI, OpenAIClientWrapper, OpenAIModels
 from sycamore.llms.gemini import Gemini, GeminiModels
 from sycamore.llms.prompts.default_prompts import SummarizeImagesJinjaPrompt
